% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2matrix.R
\name{get_chromosome}
\alias{get_chromosome}
\title{Get the chromosome of interest. 
Assume that the vcf file is sorted by chromosome.}
\usage{
get_chromosome(vcf, chromNumber)
}
\arguments{
\item{vcf}{vcf file}

\item{chromNumber}{number of the chromoseme of interest}
}
\value{
vcf object with just the chromosome chosen
}
\description{
Get the chromosome of interest. 
Assume that the vcf file is sorted by chromosome.
}
