% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_decay.R
\name{ld_decay}
\alias{ld_decay}
\title{Calculate the average LD decay for all chromosomes}
\usage{
ld_decay(vcf, max_dist = 1e+05, maf = 0.05, bin_size = 500)
}
\arguments{
\item{vcf}{a list returned by vcf2matrix function, with the genomic data, 
including positions and chromossomes of the SNPs}

\item{max_dist}{maximum distance to calculate the LD decay}

\item{maf}{minimum allele frequency to be used}

\item{bin_size}{size of the bins to calculate the decay}
}
\description{
Calculate the average LD decay for all chromosomes
}
