% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclink-package.R
\docType{package}
\name{mclink-package}
\alias{mclink}
\alias{mclink-package}
\title{The mclink package}
\description{
A package for linkage disequilibrium calculus using Monte Carlo techniques
}
\author{
\strong{Maintainer}: Kawan Santana \email{kawansantana18@gmail.com}

Authors:
\itemize{
  \item Kawan Santana \email{kawansantana18@gmail.com}
}

}
\keyword{internal}
