% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_decay.R
\name{plot.mclink_ld_decay}
\alias{plot.mclink_ld_decay}
\title{Plot LD decay}
\usage{
\method{plot}{mclink_ld_decay}(x, ...)
}
\arguments{
\item{x}{object returned by ld_decay()}

\item{...}{extended parameters}
}
\description{
Plot LD decay
}
