% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_and_process_KO_table.R
\name{read_and_process_KO_table}
\alias{read_and_process_KO_table}
\title{Read and Process KO Sample Table with Pathway Information}
\usage{
read_and_process_KO_table(in_KO_Sample_wide, pathway_infor)
}
\arguments{
\item{in_KO_Sample_wide}{Character string specifying the path to the input KO sample table file.
Should be a tab-delimited file with KO identifiers as row names.}

\item{pathway_infor}{Data frame containing pathway information, see examples.}
}
\value{
A list with log and a data frame.The data frame contains the filtered sample data, with only rows that match KOs in the
pathway information. Includes an added 'Orthology_Entry' column containing the row names.
}
\description{
This function reads a KO sample table (wide format), validates its contents against
pathway information, and returns a filtered table containing only KOs present in both datasets.
}
