% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_panmixia}
\alias{simulate_panmixia}
\title{Generate the distribuitions of the relative genetic distance observed, the distribuition of relative 
genetic distance expect and the distribuition of the standarized association index (Ia)}
\usage{
simulate_panmixia(
  data,
  iterations = 1000L,
  sample_size = 1000L,
  min_distance = 10000L
)
}
\arguments{
\item{data}{a Rcpp list with the genomic data, including positions and chromossomes of the SNPs}

\item{iterations}{number of iterations of the algorithm}

\item{sample_size}{number of SNPs to be sampled}

\item{min_distance}{minimum distance (bp) between the SNPs selected}
}
\value{
a list containing the distribuitions of the relative genetic distance observed and expect, as well the distribuition of the standarized Ia
}
\description{
Generate the distribuitions of the relative genetic distance observed, the distribuition of relative 
genetic distance expect and the distribuition of the standarized association index (Ia)
}
