% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2matrix.R
\name{vcf2matrix}
\alias{vcf2matrix}
\title{Read a vcf file and returns a sparse matrix with the genotypes}
\usage{
vcf2matrix(path, chromNumber = 0)
}
\arguments{
\item{path}{path to vcf file}

\item{chromNumber}{number of the chromosome of interest}
}
\value{
sparse matrix with the genotypes of the vcf file
}
\description{
Read a vcf file and returns a sparse matrix with the genotypes
}
