% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf_to_sparse.R
\name{vcf_to_sparse}
\alias{vcf_to_sparse}
\title{Convert a VCF file to a sparse genotype matrix}
\usage{
vcf_to_sparse(vcf_file, na_value = -1)
}
\arguments{
\item{vcf_file}{Path to the VCF file (can be .vcf or .vcf.gz).}

\item{na_value}{Numeric value used to represent missing genotypes (default
= -1).}
}
\value{
A sparse matrix of class \code{dgCMatrix} (from the \pkg{Matrix}
package), where rows are SNPs and columns are samples.
}
\description{
Reads a VCF file using the \pkg{vcfR} package, extracts the genotypes and converts them into a sparse numeric matrix where:
}
\details{
\itemize{
  \item 0 = reference allele
  \item 1 = alternative allele
  \item -1 = missing genotype
}
Each row represents a SNP and each column represents a sample (isolate).


The function accepts both haploid (0 / 1 / NA) and diploid (0/0 / 1/1 /
./.) genotype formats.
}
