\name{mclink-package}
\alias{mclink-package}
\alias{mclink}
\docType{package}
\title{
  Tools for genetic analysis using Monte Carlo methods
}
\description{
  This package allows to analysis genetic code using random sampling techniques, 
  to manipulate and understand the data through Monte Carlo methods.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Your Name, email optional.

Maintainer: Your Name <your@email.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
