\name{mclink-package}
\alias{mclink-package}
\alias{mclink}
\docType{package}
\title{\packageTitle{mclink}}
\description{\packageDescription{mclink}}
\section{Package Content}{\packageIndices{mclink}}
\author{\packageAuthor{mclink}}
\section{Maintainer}{\packageMaintainer{mclink}}
\keyword{package}
