% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2matrix.R
\name{vcf2matrix}
\alias{vcf2matrix}
\title{Read a vcf file e returns a sparse matrix equivalent of the file}
\usage{
vcf2matrix(path, chrom = 0)
}
\arguments{
\item{path}{path to vcf file}
}
\value{
sparse matrix with the genotypes
}
\description{
Read a vcf file e returns a sparse matrix equivalent of the file
}
